add test for basic markdown
