+++ markdown.go
@@ -9,6 +9,7 @@ import (
 	"regexp"
 
 	"github.com/tiechui1994/gopdf/core"
+	"encoding/json"
 )
 
 const (
@@ -18,16 +19,18 @@ const (
 )
 
 const (
-	TYPE_ORIGIN = "origin"
+	TYPE_TEXT     = "text"
+	TYPE_STRONG   = "strong"   // **strong**
+	TYPE_EM       = "em"       // *em*
+	TYPE_CODESPAN = "codespan" // `codespan`, ```codespan```
+	TYPE_CODE     = "code"     //
+	TYPE_LINK     = "link"     // [xx](http://ww)
 
-	TYPE_NORMAL = "normal"
-	TYPE_BOLD   = "bold"
-	TYPE_IALIC  = "italic"
-	TYPE_WARP   = "warp"
-	TYPE_CODE   = "code"
-	TYPE_LINK   = "link"
+	TYPE_SPACE = "space"
 
-	TYPE_HIGHLIGHT = "highlight"
+	TYPE_PARAGRAPH = "paragraph"
+	TYPE_HEADING   = "heading"
+	TYPE_LIST      = "list"
 
 	TYPE_HEADER_ONE = "1"
 	TYPE_HEADER_TWO = "2"
@@ -60,18 +63,45 @@ const (
 	defaultWarpFontSize = 10.0
 )
 
+type Token struct {
+	Type string `json:"type"`
+	Raw  string `json:"raw"`
+	Text string `json:"text"`
+
+	// list
+	Ordered bool            `json:"ordered"`
+	Start   json.RawMessage `json:"start"`
+	Loose   bool            `json:"loose"`
+	Task    bool            `json:"task"`
+	Items   []Token         `json:"items"`
+
+	// heading
+	Depth int `json:"depth"`
+
+	// link
+	Href  string `json:"href"`
+	Title string `json:"title"`
+
+	Tokens []Token `json:"tokens"`
+}
+
 type mardown interface {
 	SetText(fontFamily string, text ...string)
-	SetPadding(need bool, padding float64)
 	GenerateAtomicCell() (pagebreak, over bool, err error)
 }
 
-type combined interface {
-	AddChild(child mardown)
+type abstractMarkDown struct{}
+
+func (a *abstractMarkDown) SetText(string, ...string) {
+}
+func (a *abstractMarkDown) GenerateAtomicCell() (pagebreak, over bool, err error) {
+	return
 }
 
-// common componet
-type content struct {
+///////////////////////////////////////////////////////////////////
+
+// Atomic component
+type MdText struct {
 	Type       string
 	pdf        *core.Report
 	font       core.Font
@@ -90,22 +120,22 @@ type content struct {
 	padinglen   float64
 }
 
-func (c *content) SetText(fontFamily string, text ...string) {
+func (c *MdText) SetText(fontFamily string, text ...string) {
 	if len(text) == 0 {
 		panic("text is invalid")
 	}
 
 	var font core.Font
 	switch c.Type {
-	case TYPE_BOLD:
+	case TYPE_STRONG:
 		font = core.Font{Family: fontFamily, Size: defaultFontSize, Style: ""}
-	case TYPE_IALIC:
+	case TYPE_EM:
 		font = core.Font{Family: fontFamily, Size: defaultFontSize, Style: ""}
-	case TYPE_WARP:
+	case TYPE_CODESPAN:
 		font = core.Font{Family: fontFamily, Size: defaultWarpFontSize, Style: ""}
 	case TYPE_CODE:
 		font = core.Font{Family: fontFamily, Size: defaultWarpFontSize, Style: ""}
-	case TYPE_LINK, TYPE_NORMAL:
+	case TYPE_LINK, TYPE_TEXT:
 		font = core.Font{Family: fontFamily, Size: defaultFontSize, Style: ""}
 	default:
 		panic(fmt.Sprintf("invalid type: %v", c.Type))
@@ -134,10 +164,7 @@ func (c *content) SetText(fontFamily string, text ...string) {
 	}
 }
 
-func (c *content) SetPadding(need bool, padding float64) {
-}
-
-func (c *content) GenerateAtomicCell() (pagebreak, over bool, err error) {
+func (c *MdText) GenerateAtomicCell() (pagebreak, over bool, err error) {
 	pageEndX, pageEndY := c.pdf.GetPageEndXY()
 	x1, y := c.pdf.GetXY()
 	x2 := pageEndX
@@ -147,7 +174,7 @@ func (c *content) GenerateAtomicCell() (pagebreak, over bool, err error) {
 
 	text, width, newline := c.GetSubText(x1, x2)
 	for !c.stoped {
-		if c.Type == TYPE_WARP {
+		if c.Type == TYPE_CODESPAN {
 			// other padding is testing data
 			// bg: 248,248,255 text:1,1,1 line:	245,245,245,     Typora
 			// bg: 249,242,244 text:199,37,78 line:	245,245,245  GitLab
@@ -199,416 +226,10 @@ func (c *content) GenerateAtomicCell() (pagebreak, over bool, err error) {
 	return false, c.stoped, nil
 }
 
-func (c *content) String() string {
-	text := strings.Replace(c.remain, "\n", "|", -1)
-	return fmt.Sprintf("[type=%v,text=%v]", c.Type, text)
-}
-
-type mdimage struct {
-	pdf   *core.Report
-	image *Image
-}
-
-func (mi *mdimage) SetText(fontFamily string, filename ...string) {
-	image := NewImage(filename[0], mi.pdf)
-	mi.image = image
-}
-
-func (mi *mdimage) GenerateAtomicCell() (pagebreak, over bool, err error) {
-	mi.image.GenerateAtomicCell()
-	return
-}
-
-func (mi *mdimage) SetPadding(need bool, padding float64) {
-}
-
-// combined components
-
-type header struct {
-	Size       int
-	pdf        *core.Report
-	fonts      map[string]string
-	children   []mardown
-	lineHeight float64
-
-	text          string
-	remain        string
-	needBreakLine bool
-}
-
-func (h *header) GetFontFamily(md mardown) string {
-	switch md.(type) {
-	case *content:
-		c, _ := md.(*content)
-		switch c.Type {
-		case TYPE_NORMAL:
-			return h.fonts[FONT_NORMAL]
-		case TYPE_BOLD:
-			return h.fonts[FONT_BOLD]
-		case FONT_IALIC:
-			return h.fonts[FONT_IALIC]
-		default:
-			return h.fonts[FONT_NORMAL]
-		}
-	}
-
-	return ""
-}
-
-func (h *header) SetText(_ string, textargs ...string) {
-	var (
-		normal, code, itialic int
-	)
-	switch h.Size {
-	case HEADER_ONE:
-		normal = defaultFontSize + 12
-		h.lineHeight = defaultLineHeight + 16
-	case HEADER_TWO:
-		normal = defaultFontSize + 8
-		h.lineHeight = defaultLineHeight + 12
-	case HEADER_THR:
-		normal = defaultFontSize + 4
-		h.lineHeight = defaultLineHeight + 8
-	case HEADER_FOU:
-		normal = defaultFontSize + 3
-		h.lineHeight = defaultLineHeight + 6
-	case HEADER_FIV:
-		normal = defaultFontSize + 2
-		h.lineHeight = defaultLineHeight + 4
-	case HEADER_SIX:
-		normal = defaultFontSize
-		h.lineHeight = defaultLineHeight
-	}
-
-	_, _, _ = normal, code, itialic
-	text := textargs[0]
-
-	relink := regexp.MustCompile(`^\[(.*?)\]\((.*?)\)`)
-	reimage := regexp.MustCompile(`^\!\[image\]\((.*?)\)`)
-	runes := []rune(text)
-	n := len(runes)
-	log.Println("runes:", len(runes), "text:", strings.Replace(text, "\n", "|", -1))
-	var (
-		buf      bytes.Buffer
-		contents []mardown
-		sub      mardown  // basic composte
-		cuts     []string // mark some cut, eg, **, *, `, ```
-	)
-
-	setsub := func(m mardown) {
-		contents = append(contents, m)
-	}
-
-	defaultop := func(i *int) {
-		if buf.Len() == 0 {
-			sub = &content{pdf: h.pdf, Type: TYPE_NORMAL}
-		}
-		buf.WriteRune(runes[*i])
-		*i += 1
-	}
-
-	curIsCode := func() bool {
-		if len(cuts) == 0 {
-			return false
-		}
-
-		last := cuts[len(cuts)-1]
-		if last == "```" || last == "`" {
-			return true
-		}
-
-		return false
-	}
-
-	for i := 0; i < n; {
-		switch runes[i] {
-		case '*':
-			if len(cuts) > 0 && (cuts[len(cuts)-1] == cut_wrap || cuts[len(cuts)-1] == cut_code) {
-				defaultop(&i)
-				continue
-			}
-
-			if buf.Len() > 0 {
-				sub.SetText(h.GetFontFamily(sub), buf.String())
-				buf.Reset()
-				setsub(sub)
-			}
-
-			if i+1 < n && string(runes[i:i+2]) == cut_bold {
-				if len(cuts) == 0 || cuts[len(cuts)-1] != cut_bold {
-					sub = &content{pdf: h.pdf, Type: TYPE_BOLD}
-					cuts = append(cuts, cut_bold)
-					if runes[i+2] == '`' {
-						buf.WriteRune(runes[i+3])
-						i += 4
-					} else {
-						buf.WriteRune(runes[i+2])
-						i += 3
-					}
-
-				} else {
-					cuts = cuts[:len(cuts)-1]
-					i += 2
-				}
-				continue
-			}
-
-			if len(cuts) == 0 || cuts[len(cuts)-1] != cut_itaic {
-				sub = &content{pdf: h.pdf, Type: TYPE_IALIC}
-				cuts = append(cuts, cut_itaic)
-				if runes[i+1] == '`' {
-					buf.WriteRune(runes[i+2])
-					i += 3
-				} else {
-					buf.WriteRune(runes[i+1])
-					i += 2
-				}
-			} else {
-				cuts = cuts[:len(cuts)-1]
-				i += 1
-			}
-
-		case '`':
-			if len(cuts) > 0 && (cuts[len(cuts)-1] == cut_itaic || cuts[len(cuts)-1] == cut_bold) {
-				i += 1
-				continue
-			}
-
-			if buf.Len() > 0 {
-				sub.SetText(h.GetFontFamily(sub), buf.String())
-				buf.Reset()
-				setsub(sub)
-			}
-
-			// code text
-			if i+2 < n && string(runes[i:i+3]) == cut_code && (i == 0 || runes[i-1] == '\n') {
-				if len(cuts) == 0 || cuts[len(cuts)-1] != cut_code {
-					sub = &content{pdf: h.pdf, Type: TYPE_CODE, needpadding: true}
-					index := strings.Index(string(runes[i:]), "\n")
-					cuts = append(cuts, cut_code)
-					buf.WriteRune(runes[i+index+1])
-					i = i + index + 2
-				} else {
-					cuts = cuts[:len(cuts)-1]
-					i += 3
-				}
-				continue
-			}
-
-			// wrap
-			if i+2 < n && string(runes[i:i+3]) == cut_code {
-				if len(cuts) == 0 || cuts[len(cuts)-1] != cut_code {
-					sub = &content{pdf: h.pdf, Type: TYPE_WARP}
-					cuts = append(cuts, cut_code)
-					buf.WriteRune(runes[i+3])
-					i += 4
-				} else {
-					cuts = cuts[:len(cuts)-1]
-					i += 3
-				}
-				continue
-			}
-
-			// wrap
-			if len(cuts) == 0 || cuts[len(cuts)-1] != cut_wrap {
-				sub = &content{pdf: h.pdf, Type: TYPE_WARP}
-				cuts = append(cuts, cut_wrap)
-				buf.WriteRune(runes[i+1])
-				i += 2
-			} else {
-				cuts = cuts[:len(cuts)-1]
-				i += 1
-			}
-
-		case '!':
-			temp := string(runes[i:])
-			if !curIsCode() && reimage.MatchString(temp) {
-				if buf.Len() > 0 {
-					sub.SetText(h.GetFontFamily(sub), buf.String())
-					buf.Reset()
-					setsub(sub)
-				}
-				matchstr := reimage.FindString(temp)
-				submatch := reimage.FindStringSubmatch(temp)
-				c := &mdimage{pdf: h.pdf}
-				c.SetText("", submatch[1])
-				setsub(c)
-				i += len([]rune(matchstr))
-			} else {
-				defaultop(&i)
-			}
-
-		case '[':
-			temp := string(runes[i:])
-			if !curIsCode() && relink.MatchString(temp) {
-				if buf.Len() > 0 {
-					sub.SetText(h.GetFontFamily(sub), buf.String())
-					buf.Reset()
-					setsub(sub)
-				}
-
-				matchstr := relink.FindString(temp)
-				submatch := relink.FindStringSubmatch(temp)
-				c := &content{pdf: h.pdf, Type: TYPE_LINK}
-				c.SetText(h.fonts[FONT_NORMAL], submatch[1], submatch[2])
-				setsub(c)
-				i += len([]rune(matchstr))
-			} else {
-				defaultop(&i)
-			}
-
-		default:
-			defaultop(&i)
-		}
-	}
-
-	if buf.Len() > 0 {
-		sub.SetText(h.GetFontFamily(sub), buf.String())
-		buf.Reset()
-		setsub(sub)
-	}
-
-	h.children = contents
-}
-
-func (h *header) SetPadding(need bool, padding float64) {
-}
-
-func (h *header) GenerateAtomicCell() (pagebreak, over bool, err error) {
-	for i, comment := range h.children {
-		pagebreak, over, err = comment.GenerateAtomicCell()
-		if err != nil {
-			return
-		}
-
-		if pagebreak {
-			if over && i != len(h.children)-1 {
-				h.children = h.children[i+1:]
-				return pagebreak, len(h.children) == 0, nil
-			}
-
-			h.children = h.children[i:]
-			return pagebreak, len(h.children) == 0, nil
-		}
-	}
-	return
-}
-
-func (h *header) String() string {
-	text := strings.Replace(h.remain, "\n", "|", -1)
-	return fmt.Sprintf("[size=%v,text=%v]", h.Size, text)
-}
-
-const (
-	SORT_ORDER    = "order"
-	SORT_DISORDER = "disorder"
-)
-
-type blocksort struct {
-	pdf      *core.Report
-	font     core.Font
-	children []mardown
-
-	headerWrited bool
-	sortType     string
-	sortIndex    string
-}
-
-func (bs *blocksort) SetText(fontFamily string, _ ...string) {
-	if bs.sortType == SORT_DISORDER {
-		bs.font = core.Font{Family: fontFamily, Size: 40.0}
-	} else {
-		bs.font = core.Font{Family: fontFamily, Size: 18.0}
-	}
-
-}
-
-func (bs *blocksort) GenerateAtomicCell() (pagebreak, over bool, err error) {
-	bs.pdf.Font(bs.font.Family, bs.font.Size, bs.font.Style)
-	bs.pdf.SetFontWithStyle(bs.font.Family, bs.font.Style, bs.font.Size)
-
-	if !bs.headerWrited {
-		var text string
-		x, y := bs.pdf.GetXY()
-		switch bs.sortType {
-		case SORT_ORDER:
-			text = fmt.Sprintf(" %v. ", bs.sortIndex)
-			bs.pdf.Cell(x, y, text)
-
-		case SORT_DISORDER:
-			text = " · "
-			bs.pdf.Cell(x, y-13, text)
-		}
-
-		length := bs.pdf.MeasureTextWidth(text)
-		bs.pdf.SetXY(x+length, y)
-		bs.headerWrited = true
-	}
-
-	for i, comment := range bs.children {
-		pagebreak, over, err = comment.GenerateAtomicCell()
-		if err != nil {
-			return
-		}
-
-		if pagebreak {
-			if over && i != len(bs.children)-1 {
-				bs.children = bs.children[i+1:]
-				return pagebreak, len(bs.children) == 0, nil
-			}
-
-			bs.children = bs.children[i:]
-			return pagebreak, len(bs.children) == 0, nil
-		}
-	}
-
-	return
-}
-
-func (bs *blocksort) SetPadding(need bool, padding float64) {
-}
-
-func (bs *blocksort) AddChild(child mardown) {
-	bs.children = append(bs.children, child)
-}
-
-func (bs *blocksort) String() string {
-	var buf bytes.Buffer
-	fmt.Fprint(&buf, "(blocksort")
-	for _, child := range bs.children {
-		fmt.Fprintf(&buf, "%v", child)
-	}
-	fmt.Fprint(&buf, ")")
-
-	return buf.String()
-}
-
-//
-type blockwarp struct {
-	pdf      *core.Report
-	children []mardown
-
-	padinglen float64
-}
-
-func (bw *blockwarp) SetText(fontFamily string, _ ...string) {
-}
-
-func (bw *blockwarp) SetPadding(need bool, padding float64) {
-}
-
-func (bw *blockwarp) AddChild(child mardown) {
-	bw.children = append(bw.children, child)
-}
-
-func (bw *blockwarp) GenerateAtomicCell() (pagebreak, over bool, err error) {
-	return
-}
-
 // GetSubText, Returns the content of a string of length x2-x1.
 // This string is a substring of text.
 // After return, the remain and length will change
-func (c *content) GetSubText(x1, x2 float64) (text string, width float64, newline bool) {
+func (c *MdText) GetSubText(x1, x2 float64) (text string, width float64, newline bool) {
 	if len(c.remain) == 0 {
 		c.stoped = true
 		return "", 0, false
@@ -688,15 +309,265 @@ func (c *content) GetSubText(x1, x2 float64) (text string, width float64, newlin
 	return "", 0, false
 }
 
-// pretreatment element
-type pre struct {
-	Type  string
-	Value string
+func (c *MdText) String() string {
+	text := strings.Replace(c.remain, "\n", "|", -1)
+	return fmt.Sprintf("[type=%v,text=%v]", c.Type, text)
 }
 
-func (p *pre) String() string {
-	value := strings.Replace(p.Value, "\n", "|", -1)
-	return fmt.Sprintf("[type=%v, value=%v]", p.Type, value)
+type MdSpace struct {
+	abstractMarkDown
+	Type       string
+	pdf        *core.Report
+	lineHeight float64
+}
+
+func (c *MdSpace) GenerateAtomicCell() (pagebreak, over bool, err error) {
+	_, pageEndY := c.pdf.GetPageEndXY()
+	_, y := c.pdf.GetXY()
+
+	x, _ := c.pdf.GetPageStartXY()
+	y += c.lineHeight
+	if pageEndY-y < c.lineHeight {
+		return true, true, nil
+	}
+	c.pdf.SetXY(x, y)
+	return false, true, nil
+}
+
+type MdImage struct {
+	abstractMarkDown
+	pdf   *core.Report
+	image *Image
+}
+
+func (mi *MdImage) SetText(fontFamily string, filename ...string) {
+	image := NewImage(filename[0], mi.pdf)
+	mi.image = image
+}
+
+func (mi *MdImage) GenerateAtomicCell() (pagebreak, over bool, err error) {
+	mi.image.GenerateAtomicCell()
+	return
+}
+
+// Combination components
+
+type MdHeader struct {
+	Size       int
+	pdf        *core.Report
+	fonts      map[string]string
+	children   []mardown
+	lineHeight float64
+
+	text          string
+	remain        string
+	needBreakLine bool
+}
+
+func (h *MdHeader) GetFontFamily(md mardown) string {
+	switch md.(type) {
+	case *MdText:
+		c, _ := md.(*MdText)
+		switch c.Type {
+		case TYPE_TEXT:
+			return h.fonts[FONT_NORMAL]
+		case TYPE_STRONG:
+			return h.fonts[FONT_BOLD]
+		case TYPE_EM:
+			return h.fonts[FONT_IALIC]
+		default:
+			return h.fonts[FONT_NORMAL]
+		}
+	}
+
+	return ""
+}
+
+func (h *MdHeader) GenerateAtomicCell() (pagebreak, over bool, err error) {
+	for i, comment := range h.children {
+		pagebreak, over, err = comment.GenerateAtomicCell()
+		if err != nil {
+			return
+		}
+
+		if pagebreak {
+			if over && i != len(h.children)-1 {
+				h.children = h.children[i+1:]
+				return pagebreak, len(h.children) == 0, nil
+			}
+
+			h.children = h.children[i:]
+			return pagebreak, len(h.children) == 0, nil
+		}
+	}
+	return
+}
+
+func (h *MdHeader) String() string {
+	text := strings.Replace(h.remain, "\n", "|", -1)
+	return fmt.Sprintf("[size=%v,text=%v]", h.Size, text)
+}
+
+type MdParagraph struct {
+	abstractMarkDown
+	pdf      *core.Report
+	fonts    map[string]string
+	children []mardown
+	Type     string
+}
+
+func (p *MdParagraph) SetToken(token Token) error {
+	if p.fonts == nil || len(p.fonts) == 0 {
+		return fmt.Errorf("no fonts")
+	}
+	if token.Type != TYPE_PARAGRAPH {
+		return fmt.Errorf("invalid type")
+	}
+
+	for _, tok := range token.Tokens {
+		var child mardown
+		switch tok.Type {
+		case TYPE_LINK:
+			child = &MdText{
+				pdf:  p.pdf,
+				Type: TYPE_LINK,
+			}
+			child.SetText(p.fonts[FONT_NORMAL], tok.Text, tok.Href)
+		case TYPE_TEXT:
+			child = &MdText{
+				pdf:  p.pdf,
+				Type: TYPE_TEXT,
+			}
+			child.SetText(p.fonts[FONT_NORMAL], tok.Text)
+		case TYPE_EM:
+			child = &MdText{
+				pdf:  p.pdf,
+				Type: TYPE_EM,
+			}
+			child.SetText(p.fonts[FONT_IALIC], tok.Text)
+		case TYPE_CODESPAN:
+			child = &MdText{
+				pdf:  p.pdf,
+				Type: TYPE_CODESPAN,
+			}
+			child.SetText(p.fonts[FONT_NORMAL], tok.Text)
+		case TYPE_CODE:
+			child = &MdText{
+				pdf:  p.pdf,
+				Type: TYPE_CODE,
+			}
+			child.SetText(p.fonts[FONT_NORMAL], tok.Text)
+		case TYPE_STRONG:
+			child = &MdText{
+				pdf:  p.pdf,
+				Type: TYPE_STRONG,
+			}
+			child.SetText(p.fonts[FONT_BOLD], tok.Text)
+		}
+
+		p.children = append(p.children, child)
+	}
+
+	return nil
+}
+
+func (p *MdParagraph) GenerateAtomicCell() (pagebreak, over bool, err error) {
+	for i, comment := range p.children {
+		pagebreak, over, err = comment.GenerateAtomicCell()
+		if err != nil {
+			return
+		}
+
+		if pagebreak {
+			if over && i != len(p.children)-1 {
+				p.children = p.children[i+1:]
+				return pagebreak, len(p.children) == 0, nil
+			}
+
+			p.children = p.children[i:]
+			return pagebreak, len(p.children) == 0, nil
+		}
+	}
+	return
+}
+
+///////////////////////////////////////////////////////
+
+const (
+	SORT_ORDER    = "order"
+	SORT_DISORDER = "disorder"
+)
+
+type blocksort struct {
+	pdf      *core.Report
+	font     core.Font
+	children []mardown
+
+	headerWrited bool
+	sortType     string
+	sortIndex    string
+}
+
+func (bs *blocksort) SetText(fontFamily string, _ ...string) {
+	if bs.sortType == SORT_DISORDER {
+		bs.font = core.Font{Family: fontFamily, Size: 40.0}
+	} else {
+		bs.font = core.Font{Family: fontFamily, Size: 18.0}
+	}
+
+}
+
+func (bs *blocksort) GenerateAtomicCell() (pagebreak, over bool, err error) {
+	bs.pdf.Font(bs.font.Family, bs.font.Size, bs.font.Style)
+	bs.pdf.SetFontWithStyle(bs.font.Family, bs.font.Style, bs.font.Size)
+
+	if !bs.headerWrited {
+		var text string
+		x, y := bs.pdf.GetXY()
+		switch bs.sortType {
+		case SORT_ORDER:
+			text = fmt.Sprintf(" %v. ", bs.sortIndex)
+			bs.pdf.Cell(x, y, text)
+
+		case SORT_DISORDER:
+			text = " · "
+			bs.pdf.Cell(x, y-13, text)
+		}
+
+		length := bs.pdf.MeasureTextWidth(text)
+		bs.pdf.SetXY(x+length, y)
+		bs.headerWrited = true
+	}
+
+	for i, comment := range bs.children {
+		pagebreak, over, err = comment.GenerateAtomicCell()
+		if err != nil {
+			return
+		}
+
+		if pagebreak {
+			if over && i != len(bs.children)-1 {
+				bs.children = bs.children[i+1:]
+				return pagebreak, len(bs.children) == 0, nil
+			}
+
+			bs.children = bs.children[i:]
+			return pagebreak, len(bs.children) == 0, nil
+		}
+	}
+
+	return
+}
+
+func (bs *blocksort) String() string {
+	var buf bytes.Buffer
+	fmt.Fprint(&buf, "(blocksort")
+	for _, child := range bs.children {
+		fmt.Fprintf(&buf, "%v", child)
+	}
+	fmt.Fprint(&buf, ")")
+
+	return buf.String()
 }
 
 type MarkdownText struct {
@@ -727,449 +598,70 @@ func NewMarkdownText(pdf *core.Report, x float64, fonts map[string]string) (*Mar
 	return &mt, nil
 }
 
-func (mt *MarkdownText) GetFontFamily(md mardown) string {
-	switch md.(type) {
-	case *content:
-		c, _ := md.(*content)
-		switch c.Type {
-		case TYPE_NORMAL:
-			return mt.fonts[FONT_NORMAL]
-		case TYPE_BOLD:
-			return mt.fonts[FONT_BOLD]
-		case FONT_IALIC:
-			return mt.fonts[FONT_IALIC]
-		default:
-			return mt.fonts[FONT_NORMAL]
-		}
-	}
-
-	return ""
-}
-
-const (
-	cut_code = "```"
-	cut_wrap = "`"
-
-	cut_bold  = "**"
-	cut_itaic = "*"
-)
-
-var (
-	// [\n \t=#%@&"':<>,(){}_;/\?\.\+\-\=\^\$\[\]\!]
-	rerest = regexp.MustCompile(`^\n[\t ]*?\n(\n[\t ]*?\n)*`)
-
-	// ^\*{2}\S
-	// ((\S+\s|\s\S+|\S)+\n)*
-	// \S+\*{2}
-	// need handle "\n\s*\n" special condition
-	reboldprefix = regexp.MustCompile(`^\*{2}\S`)
-	reboldsuffix = regexp.MustCompile(`\S\*{2}$`)
-
-	// \*\S
-	// ((\S+\s|\s\S+|\S)+\n)*
-	// \S+\*
-	// need handle "\n\s*\n" special condition
-	reitalic = regexp.MustCompile(`^\*\S((\S+\s|\s\S+|\S)+\n)*\S+\*`)
-
-	reitalicprefix = regexp.MustCompile(`^\*\S`)
-	reitalicsuffix = regexp.MustCompile(`\S\*$`)
-
-	// need handle "\n\s*\n" special condition
-	redel = regexp.MustCompile(`^~{2}\S((\S+\s|\s\S+|\S)+\n)*\S+~{2}`)
-
-	recodeprefix     = regexp.MustCompile("^`{3}.*?\n")
-	recodewarp       = regexp.MustCompile("^`{3}.*`{3}")
-	retextwarpprefix = regexp.MustCompile("^`")
-
-	// when current is '='
-	// first global search
-	rehighlight = regexp.MustCompile(`.*\n\s{0,3}=+$`)
-
-	// remove \n\s*\n
-	relink  = regexp.MustCompile(`^\[.*\n?(.*\n|.*)\]\(\n?(.*)\n?\)`)
-	reimage = regexp.MustCompile(`^\!\[image\]\(\n?(.*)\n?\)`)
-
-	// stoped by \n\s*\n
-	rensort = regexp.MustCompile(`^-\s{1,4}((\S+\s|\s\S+|\S)+\n)*`)
-
-	reheader = regexp.MustCompile(`^(#)+.*`)
-
-	// need handle "\n\s*\n" special condition
-	reref = regexp.MustCompile(`^>((\S+\s|\s\S+|\S)+\n)*`)
-)
-
-func (mt *MarkdownText) parseBoldText(i int) (ok bool, ) {
-	return
-}
-
-func (mt *MarkdownText) SetText(text string) *MarkdownText {
-	var (
-		contents []mardown
-	)
-	pres := mt.PreProccesText(text)
-	for _, pre := range pres {
-		log.Printf("%v", pre.String())
-		switch pre.Type {
-		case TYPE_WARP, TYPE_CODE, TYPE_BOLD, TYPE_NORMAL:
-			c := content{
+func (mt *MarkdownText) SetTokens(tokens []Token) {
+	for _, token := range tokens {
+		switch token.Type {
+		case TYPE_PARAGRAPH:
+			paragraph := &MdParagraph{
+				pdf:   mt.pdf,
+				fonts: mt.fonts,
+				Type:  token.Type,
+			}
+			paragraph.SetToken(token)
+			mt.contents = append(mt.contents, paragraph)
+		case TYPE_SPACE:
+			space := &MdSpace{pdf: mt.pdf, Type: token.Type, lineHeight: defaultLineHeight}
+			mt.contents = append(mt.contents, space)
+		case TYPE_LINK:
+			link := &MdText{
 				pdf:  mt.pdf,
-				Type: pre.Type,
+				Type: TYPE_LINK,
 			}
-			c.SetText(mt.GetFontFamily(&c), pre.Value)
-			contents = append(contents, &c)
-		}
-
-	}
-
-	mt.contents = contents
-
-	return mt
-}
-
-func (mt *MarkdownText) GetWritedLines() int {
-	return mt.writedLines
-}
-
-func (mt *MarkdownText) PreProccesText(text string) []pre {
-	/*
-	// header and line
-	# xx
-	## xx
-
-	// bold and line
-	xxx
-	==
-
-	// del line
-	~~ xx ~~
-	*/
-
-	// when header (#), the '\n' can not replace
-	// when code (```) include '\n', code text can not replace
-	// when hence (`), if contains '\n[ ]*\n', not hence, not replace, else replace '\n' with ' '
-	// when link ([]()), if [] contains '\n[ ]*\n', not link, if () contains more than two lines,
-	// not link. else replace '\n' with ' '. and image is so on
-	// when sort (-), if contains '\n[ ]+\n', after must new line, else replace '\n' with ' '
-	// when wrap (*), if after|before * is ' ' or '\n[ ]*\n' not warp
-
-	runes := []rune(text)
-	n := len(runes)
-
-	// break some 
-	rebreak := regexp.MustCompile(`\n\s*\n`)
-
-	type prehighlight struct {
-		pre
-		last int
-	}
-
-	var (
-		i           int
-		buf         bytes.Buffer
-		pres        []pre
-		ignoreindex map[int]prehighlight
-	)
-	ignoreindex = make(map[int]prehighlight)
-	hgindexs := rehighlight.FindAllStringIndex(text, -1)
-	if len(hgindexs) > 0 {
-		for _, tuple := range hgindexs {
-			index, last := tuple[0], tuple[1]
-			brindex := strings.LastIndex(string(runes[index:last]), "\n")
-			ignoreindex[index] = prehighlight{
-				pre: pre{
-					Type:  TYPE_HIGHLIGHT,
-					Value: string(runes[index:brindex]),
-				},
-				last: last,
+			link.SetText(mt.fonts[FONT_NORMAL], token.Text, token.Href)
+			mt.contents = append(mt.contents, link)
+		case TYPE_TEXT:
+			text := &MdText{
+				pdf:  mt.pdf,
+				Type: TYPE_TEXT,
 			}
+			text.SetText(mt.fonts[FONT_NORMAL], token.Text)
+			mt.contents = append(mt.contents, text)
+		case TYPE_EM:
+			em := &MdText{
+				pdf:  mt.pdf,
+				Type: TYPE_EM,
+			}
+			em.SetText(mt.fonts[FONT_IALIC], token.Text)
+			mt.contents = append(mt.contents, em)
+		case TYPE_CODESPAN:
+			codespan := &MdText{
+				pdf:  mt.pdf,
+				Type: TYPE_CODESPAN,
+			}
+			codespan.SetText(mt.fonts[FONT_NORMAL], token.Text)
+			mt.contents = append(mt.contents, codespan)
+		case TYPE_CODE:
+			code := &MdText{
+				pdf:  mt.pdf,
+				Type: TYPE_CODE,
+			}
+			code.SetText(mt.fonts[FONT_NORMAL], token.Text+"\n")
+			mt.contents = append(mt.contents, code)
+		case TYPE_STRONG:
+			strong := &MdText{
+				pdf:  mt.pdf,
+				Type: TYPE_STRONG,
+			}
+			strong.SetText(mt.fonts[FONT_BOLD], token.Text)
+			mt.contents = append(mt.contents, strong)
 		}
 	}
-	resetbuf := func() {
-		if buf.Len() > 0 {
-			pres = append(pres, pre{
-				Type:  TYPE_NORMAL,
-				Value: buf.String(),
-			})
-			buf.Reset()
-		}
-	}
-	defaultop := func() {
-		temp := string(runes[i:])
-		rebreakseg := regexp.MustCompile(`^\n\s*\n`)
-
-		if runes[i] == '\n' && rebreakseg.MatchString(temp) {
-			buf.WriteRune('\n')
-			resetbuf()
-			matched := rebreakseg.FindString(temp)
-			i += len([]rune(matched))
-			return
-		}
-
-		buf.WriteRune(runes[i])
-		i++
-	}
-
-	for i < n {
-		if val, ok := ignoreindex[i]; ok {
-			pres = append(pres, val.pre)
-			i = val.last
-			continue
-		}
-
-		temp := string(runes[i:])
-		switch runes[i] {
-		case '*':
-			log.Println("--------->", string(temp[:8]))
-			if reboldprefix.MatchString(temp) {
-				resetbuf()
-				k := i + 2
-				ok, index := deepinSearch(runes[k:], "**", true)
-				log.Println(ok, string(runes[i:k+index]), index)
-				if ok && reboldsuffix.MatchString(string(runes[k:k+index])) {
-					pres = append(pres, pre{
-						Type:  TYPE_BOLD,
-						Value: strings.Trim(string(runes[i:k+index]), "**"),
-					})
-					i = k + index
-					continue
-				}
-
-				pres = append(pres, pre{
-					Type:  TYPE_ORIGIN,
-					Value: string(runes[i:k+index]),
-				})
-				i = k + index
-				continue
-			}
-
-			if reitalicprefix.MatchString(temp) {
-				resetbuf()
-				matched := reitalic.FindString(temp)
-				i += len([]rune(matched))
-				if rebreak.MatchString(matched) {
-					matched = rebreak.ReplaceAllString(matched, "\n")
-					pres = append(pres, pre{
-						Type:  TYPE_ORIGIN,
-						Value: strings.Replace(matched, "\n", " ", -1),
-					})
-					continue
-				}
-
-				matched = strings.Replace(matched, "\n", " ", -1)
-				pres = append(pres, pre{
-					Type:  TYPE_IALIC,
-					Value: strings.Trim(matched, "*"),
-				})
-				continue
-			}
-
-			defaultop()
-
-		case '`':
-			if (i == 0 || runes[i-1] == '\n') && recodeprefix.MatchString(temp) {
-				resetbuf()
-				prefix := recodeprefix.FindString(temp)
-				k := i + len([]rune(prefix))
-				ok, index := deepinSearch(runes[k:], "\n```", false)
-				if ok {
-					pres = append(pres, pre{
-						Type:  TYPE_CODE,
-						Value: strings.Trim(string(runes[k:k+index]), "```"),
-					})
-					i = k + index
-					continue
-				}
-
-				pres = append(pres, pre{
-					Type:  TYPE_ORIGIN,
-					Value: string(runes[i:k+index]),
-				})
-				i = k + index
-				continue
-			} else if recodewarp.MatchString(temp) {
-				resetbuf()
-				matched := recodewarp.FindString(temp)
-				i += len([]rune(matched))
-				pres = append(pres, pre{
-					Type:  TYPE_WARP,
-					Value: strings.Trim(matched, "````"),
-				})
-				continue
-			} else if retextwarpprefix.MatchString(temp) {
-				resetbuf()
-				prefix := retextwarpprefix.FindString(temp)
-				k := i + len([]rune(prefix))
-				ok, index := deepinSearch(runes[k:], "`", true)
-				if ok {
-					pres = append(pres, pre{
-						Type:  TYPE_WARP,
-						Value: strings.Trim(string(runes[i:k+index]), "`"),
-					})
-					i = k + index
-					continue
-				}
-
-				pres = append(pres, pre{
-					Type:  TYPE_ORIGIN,
-					Value: string(runes[i:k+index]),
-				})
-				i = k + index
-				continue
-			}
-
-			defaultop()
-
-		case '#':
-			if (i == 0 || runes[i-1] == '\n') && reheader.MatchString(temp) {
-				resetbuf()
-				matched := reheader.FindString(temp)
-				i += len([]rune(matched))
-				var size int
-				matched = strings.TrimLeftFunc(matched, func(r rune) bool {
-					if size == HEADER_SIX {
-						return false
-					}
-					if r == '#' {
-						size++
-					}
-					return r == '#'
-				})
-				pres = append(pres, pre{
-					Type:  fmt.Sprintf("%v", size),
-					Value: matched,
-				})
-				continue
-			}
-			defaultop()
-
-		case '!':
-			if reimage.MatchString(temp) {
-				resetbuf()
-				matched := reimage.FindString(temp)
-				i += len([]rune(matched))
-				if rebreak.MatchString(matched) {
-					matched = rebreak.ReplaceAllString(matched, "\n")
-					pres = append(pres, pre{
-						Type:  TYPE_ORIGIN,
-						Value: strings.Replace(matched, "\n", " ", -1),
-					})
-					continue
-				}
-
-				si := strings.Index(matched, "]")
-				se := strings.LastIndex(matched, ")")
-				pres = append(pres, pre{
-					Type:  TYPE_IMAGE,
-					Value: matched[si+2:se],
-				})
-				continue
-			}
-			defaultop()
-
-		case '[':
-			if relink.MatchString(temp) {
-				resetbuf()
-				matched := relink.FindString(temp)
-				i += len([]rune(matched))
-				if rebreak.MatchString(matched) {
-					matched = rebreak.ReplaceAllString(matched, "\n")
-					pres = append(pres, pre{
-						Type:  TYPE_ORIGIN,
-						Value: strings.Replace(matched, "\n", " ", -1),
-					})
-					continue
-				}
-
-				matched = strings.Replace(matched, "\n", " ", -1)
-				pres = append(pres, pre{
-					Type:  TYPE_LINK,
-					Value: matched,
-				})
-				continue
-			}
-			defaultop()
-
-		case '~':
-			if redel.MatchString(temp) {
-				resetbuf()
-				matched := redel.FindString(temp)
-				i += len([]rune(matched))
-				if rebreak.MatchString(matched) {
-					matched = rebreak.ReplaceAllString(matched, "\n")
-					pres = append(pres, pre{
-						Type:  TYPE_ORIGIN,
-						Value: strings.Replace(matched, "\n", " ", -1),
-					})
-					continue
-				}
-
-				matched = strings.Replace(matched, "\n", " ", -1)
-				pres = append(pres, pre{
-					Type:  TYPE_DELETE,
-					Value: strings.Trim(matched, "~~"),
-				})
-				continue
-			}
-
-		case '-':
-			if (i == 0 || runes[i-1] == '\n') && rensort.MatchString(temp) {
-				resetbuf()
-				matched := rensort.FindString(temp)
-
-				// break
-				if rebreak.MatchString(matched) {
-					index := rebreak.FindAllStringIndex(matched, 1)
-					matched = matched[:index[0][0]+1]
-				}
-
-				i += len([]rune(matched))
-				pres = append(pres, pre{
-					Type:  TYPE_NSORT,
-					Value: strings.Replace(matched, "\n", " ", -1),
-				})
-				continue
-			}
-			defaultop()
-
-		case '>':
-			if (i == 0 || runes[i-1] == '\n') && reref.MatchString(temp) {
-				resetbuf()
-				matched := reref.FindString(temp)
-				if rebreak.MatchString(matched) {
-					index := rebreak.FindAllStringIndex(matched, 1)
-					matched = matched[:index[0][0]+1]
-				}
-
-				i += len([]rune(matched))
-
-				rereplace := regexp.MustCompile(`\n\s*>`)
-				matched = rereplace.ReplaceAllString(matched, " ")
-				matched = strings.Replace(matched, "\n", " ", -1)
-				pres = append(pres, pre{
-					Type:  TYPE_REFER,
-					Value: matched,
-				})
-				continue
-			}
-			defaultop()
-
-		default:
-			defaultop()
-		}
-	}
-
-	resetbuf()
-
-	return pres
 }
 
 func (mt *MarkdownText) GenerateAtomicCell() (err error) {
 	if len(mt.contents) == 0 {
 		return fmt.Errorf("not set text")
 	}
-
-	log.Println()
-	log.Println("++++++++++++++++++++++++++++++++++++++++++++++++++++")
 
 	for i := 0; i < len(mt.contents); {
 		content := mt.contents[i]
@@ -1196,39 +688,4 @@ func (mt *MarkdownText) GenerateAtomicCell() (err error) {
 	}
 
 	return nil
-}
-
-func deepinSearch(runes []rune, cut string, igblankline bool) (ok bool, index int) {
-	cs := len([]rune(cut))
-	var (
-		i        int
-		cur, pre int
-	)
-
-	cur, pre = 0, -1
-	for i < len(runes) {
-		if igblankline {
-			//  \f\n\r\t\v
-			if runes[i] != ' ' && runes[i] != '\f' && runes[i] != '\r' && runes[i] != '\t' && runes[i] != '\v' {
-				cur++
-			}
-
-			if runes[i] == '\n' {
-				if cur == 0 && pre == 0 {
-					return false, i
-				}
-				pre, cur = cur, 0
-				i++
-				continue
-			}
-		}
-
-		if i+cs < len(runes) && string(runes[i:i+cs]) == cut {
-			return true, i + cs
-		}
-
-		i++
-	}
-
-	return false, cs
 }
