--- markdown/hello.json
@@ -1,17 +1,0 @@
-{
-  "newline": "^\n+/,",
-  "code":"^( {4}[^\n]+\n*)+",
-  "fences": "^ {0,3}(`{3,}(?=[^`\n]*\n)|~{3,})([^\n]*)\n(?:|([\s\S]*?)\n)(?: {0,3}\1[~`]* *(?:\n+|$)|$)",
-  "hr": "^ {0,3}((?:- *){3,}|(?:_ *){3,}|(?:\* *){3,})(?:\n+|$)",
-  "heading":"^ {0,3}(#{1,6}) +([^\n]*?)(?: +#+)? *(?:\n+|$)",
-  "blockquote":"^( {0,3}> ?(paragraph|[^\n]*)(?:\n|$))+",
-  "list": "^( {0,3})(bull) [\s\S]+?(?:hr|def|\n{2,}(?! )(?!\1bull )\n*|\s*$)",
-  "html": "^ {0,3}(?:<(script|pre|style)[\s>][\s\S]*?(?:</\1>[^\n]*\n+|$)|comment[^\n]*(\n+|$)|<\?[\s\S]*?\?>\n*|<![A-Z][\s\S]*?>\n*|<!\[CDATA\[[\s\S]*?\]\]>\n*|</?(tag)(?: +|\n|/?>)[\s\S]*?(?:\n{2,}|$)|<(?!script|pre|style)([a-z][\w-]*)(?:attribute)*? */?>(?=[ \t]*(?:\n|$))[\s\S]*?(?:\n{2,}|$)|</(?!script|pre|style)[a-z][\w-]*\s*>(?=[ \t]*(?:\n|$))[\s\S]*?(?:\n{2,}|$))",
-  "def": "^ {0,3}\[(label)\]: *\n? *<?([^\s>]+)>?(?:(?: +\n? *| *\n *)(title))? *(?:\n+|$)"
-  "lheading": "^([^\n]+)\n {0,3}(=+|-+) *(?:\n+|$)",
-  "_paragraph":"^([^\n]+(?:\n(?!hr|heading|lheading|blockquote|fences|list|html)[^\n]+)*)",
-  "text":"^[^\n]+",
-  "_label":"(?!\s*\])(?:\\[\[\]]|[^\[\]])+",
-  "_title": "(?:"(?:\\"?|[^"\\])*"|'[^'\n]*(?:\n[^'\n]+)*\n?'|\([^()]*\))"
-}
-
