--- markdown/rule.go
@@ -1,12 +1,0 @@
-package markdown
-
-type Rule struct {
-	block struct {
-	}
-	inline struct {
-	}
-}
-
-func GetDefaultRules() *Rule {
-	return &Rule{}
-}
